update de lo útlimo.

+++ similitud/c_sim_temporal_series_euc.cpp
@@ -6,7 +6,7 @@
 using namespace std;
 
 double calc_euclidean(vector<int> actual, vector<int> pred){
-    long acum = 0;
+    long long acum = 0;
     for(int i=0; i<actual.size(); i++){
         long diff = actual[i] - pred[i];
         diff *= diff;
@@ -18,7 +18,7 @@ double calc_euclidean(vector<int> actual, vector<int> pred){
 double dist_euclidea(int a1, int b1, int a2, int b2){
     int a = a1 - a2;
     int b = b1 - b2;
-    int c = a*a + b*b;
+    long long c = a*a + b*b;
     return sqrt(c);
 }
